backend/pyproject.toml
[tool.poetry]
name = "web-reader-backend"
version = "0.1.0"
description = "Backend API for the Web Reader project (FastAPI service)."
authors = ["Web Reader Developers <devnull@example.com>"]
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.121"
uvicorn = { version = "^0.38.0", extras = ["standard"] }
python-multipart = "^0.0.20"
websockets = "^15.0"
httpx = "^0.28.0"
pydantic = "^2.10.0"
loguru = "^0.7.2"
aiofiles = "^25.1"

[tool.poetry.group.test.dependencies]
pytest = "^9.0.1"
pytest-asyncio = "^1.3"
pytest-cov = "^7.0.0"
pytest-timeout = "^2.2.0"
pytest-mock = "^3.14.0"
ruff = "^0.14"
python-dotenv = "^1.2"

[tool.poetry.group.debug.dependencies]
debugpy = "^1.8"
watchfiles = "^1.1"

[tool.poetry.group.dev.dependencies]
ruff = "^0.14"

[build-system]
requires = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"
